Backend/requirements.txt
flask
flask-cors
mysql-connector-python
numpy
pandas
scikit-learn